+++ packages/db/src/index.ts
@@ -139,6 +139,12 @@ export {
 export {
   QuerySearchClient,
 } from './lib/QuerySearchClient';
+export {
+  BaseSearch,
+} from './search/BaseSearch';
+export {
+  ExecutionSearch,
+} from './search/ExecutionSearch';
 
 export { AsyncOperationPgModel } from './models/async_operation';
 export { BasePgModel } from './models/base';